mostrar habitos implementado

+++ app/src/main/java/com/example/controla_tus_habitos/model/entities_pojos/audio/Audio.java
@@ -2,11 +2,11 @@ package com.example.controla_tus_habitos.model.entities_pojos.audio;
 
 public class Audio {
     private long id;
-    private int habitoId;
+    private long habitoId;
     private String audioPath;
 
     // Constructor
-    public Audio(long id, int habitoId, String audioPath) {
+    public Audio(long id, long habitoId, String audioPath) {
         this.id = id;
         this.habitoId = habitoId;
         this.audioPath = audioPath;
@@ -21,11 +21,11 @@ public class Audio {
         this.id = id;
     }
 
-    public int getHabitoId() {
+    public long getHabitoId() {
         return habitoId;
     }
 
-    public void setHabitoId(int habitoId) {
+    public void setHabitoId(long habitoId) {
         this.habitoId = habitoId;
     }
